% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fragCont.r
\name{.mostCommonClass}
\alias{.mostCommonClass}
\title{Calculates the most common fragmentation class across a set of cells}
\usage{
.mostCommonClass(x)
}
\description{
Calculates the most common fragmentation class across a set of cells
}
