% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fragCont.r
\name{.mostCommonClassFreq}
\alias{.mostCommonClassFreq}
\title{Calculates the average frequency of the most common fragmentation class across a set of cells}
\usage{
.mostCommonClassFreq(x)
}
\description{
Calculates the average frequency of the most common fragmentation class across a set of cells
}
