% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fragmentation.R
\name{frag}
\alias{frag}
\title{Fragmentation}
\usage{
frag(counts, weartime = NULL, thresh.lower = 100, bout.length = 1, ...)
}
\arguments{
\item{counts}{Matrix of multiple (or 1) rows with 1440 columns.}

\item{weartime}{Matrix of same dimension as \code{counts} and must
consist of 1's and 0's.  If unspecified, assumed to be when
\code{counts > 0}.}

\item{thresh.lower}{Lower cut-off for count values in intensity range.
Passed to \code{\link{accel.bouts}}}

\item{bout.length}{minimum length of an activity bout.
Passed to \code{\link{accel.bouts}}}

\item{...}{additional arguments passed to  \code{\link{accel.bouts}}}
}
\value{
List of mean rest bout, mean activity bout, the total data,
and the reciprocal of the bouts
}
\description{
Fragmentation
}
\examples{
counts = matrix(rpois(1440*5, lambda = 5), ncol = 1440)
res = frag(counts, thresh.lower = 2)

counts = example_activity_data$counts
weartime = example_activity_data$weartime
id_to_run = counts$ID[1]
counts = counts[ counts$ID == id_to_run, ]
counts$ID = counts$visit = NULL
weartime = weartime[ weartime$ID == id_to_run, ]
weartime$ID = weartime$visit = NULL

res = frag(counts, weartime = weartime, thresh.lower = 100)

}
