% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fragmentation.r
\docType{package}
\name{fragmentation}
\alias{fragmentation}
\title{fragmentation: Forest fragmentation}
\description{
Calculate forest (or similar) fragmentation connectivity, density, and fragmentation class as per \href{https://www.jstor.org/stable/26271763}{Riitters et al. 2020} \emph{Conservation Ecology} 4:3. (also see \href{https://www.ecologyandsociety.org/vol4/iss2/art3/errata/january26.2001.html}{errata}). Forest density, connectivity, and fragmentation indices can be calculated for binary (forest/non-forest) and continuous rasters.
}
\details{
\code{\link{fragBinary}}: Calculate density, connectivity, and fragmentatation class for rasters with binary values. \cr
		\code{\link{fragCont}}: Calculate density, connectivity, and fragmentatation class for rasters with continuous values. \cr
		\code{\link{rocForest}}: Example raster with percent forest cover data. \cr
}
\author{
Adam B. Smith
}
