% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/long_frag.R
\name{multi_frag}
\alias{multi_frag}
\title{Fragmentation of Mulitple Subjects}
\usage{
multi_frag(counts, weartime, id = "ID", visit = "visit",
  thresh.lower = 100, bout.length = 1)
}
\arguments{
\item{counts}{A data.frame of multiple (or 1) rows with 1441 or 1442 columns.}

\item{weartime}{data.frame of same dimension as \code{counts} and must
consist of 1's and 0's.  If unspecified, assumed to be when
\code{counts > 0}.}

\item{id}{column name/variable in \code{counts} and \code{weartime}
corresponding to an ID.}

\item{visit}{column name/variable in \code{counts} and \code{weartime}
 corresponding to a visit.
If not only one visit per \code{id}, set this to \code{NULL}}

\item{thresh.lower}{Lower cut-off for count values in intensity range.
Passed to \code{\link{accel.bouts}}}

\item{bout.length}{minimum length of an activity bout.
Passed to \code{\link{accel.bouts}}}

\item{...}{additional arguments passed to  \code{\link{accel.bouts}}}
}
\value{
List of mean rest bout, mean activity bout, the total data,
and the reciprocal of the bouts
}
\description{
Fragmentation of Mulitple Subjects
}
\examples{
counts = example_activity_data$counts
weartime = example_activity_data$weartime
res = multi_frag(counts, weartime = weartime, thresh.lower = 100)
}
