% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rocForest.r
\docType{data}
\name{rocForest}
\alias{rocForest}
\title{Percent forest cover in year 2000 for a portion of the Republic of the Congo}
\format{
\code{'SpatRaster'}.
}
\description{
Raster of occurrence/non-occurrence of forest cover in a portion of the Republic of the Congo. Cells are 30-m in resolution. Values represent percent forest cover.
}
\references{
Hansen, M.C., Potapov, P.V., Moore, R., Hancher, M., Turubanova, S.A. Tyukavina, A., Thau, D., Stehman, S.V., Goetz, S.J., Loveland, T.R., Kommareddy, A., Egerov A., Chini, L., Justice, C.O., and Townshend, J.R.G.  2013.  High-resolution global maps of 21st-century forest cover change. \emph{Science} 342:850-853. \doi{10.1126/science.1244693}.
}
