% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/long_frag.R
\name{wide_to_long}
\alias{wide_to_long}
\title{Convert Wide 1440+ data to Long}
\usage{
wide_to_long(x, id = "ID", visit = NULL)
}
\arguments{
\item{x}{A \code{data.frame} of and ID variable, 1440 measurements, and
an optional visit variable}

\item{id}{column name/variable in \code{x} corresponding to an ID}

\item{visit}{column name/variable in \code{x} corresponding to a visit.
If not only one visit per \code{id}, set this to \code{NULL}}
}
\value{
A \code{data.frame} of long values
}
\description{
Convert Wide 1440+ data to Long
}
\examples{
long_counts = wide_to_long(example_activity_data$counts, id = "ID",
visit = "visit")
}
